crawlers/gringo/pyproject.toml
[tool.poetry]
name = "gringo-crawler"
version = "0.1.0"
description = "Crawler for gringo.co.il website"
authors = ["Noam"]

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
beautifulsoup4 = "^4.12.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
lxml = "^4.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
